cf/requirements.txt
requests==2.25.1
google-cloud-secret-manager==2.2.0
google-cloud-securitycenter==1.1.0